+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - chore: add a visible button to exit grand mode
 - chore: improve axis drag handle hover indication from panning
+- chore: auto-fade out legend sidebar in grand tour mode
 
 ## v0.4.1